Combine the Engagement Actions into a single action where possible
